Isse113: Refactoring InformationSheetContent file: Commit 1: 30% work done
